Made Minigame Class
